[Test] drop manifest-mirror tests covered by validate_manifest C3/C4/C5
PR #1292 added strict-parity gates C3 (ctor defaults + kw-only), C4
(forward positional names), and C5 (Slot S12+S13 dispatch_kernel
routing) to scripts/validate_manifest.py. Per-op pytest fanouts that
mirror the manifest are now redundant.

- Delete tests/ops/test_op_base_invariants.py (191 nodes). Both the
  static __init__ AST invariant and the runtime kernel_map override
  smoke check internal scaffolding, not external op surface; C5 is
  the canonical enforcer.

- Trim tests/ops/test_elementwise_binary_alignment.py: drop
  test_manifest_entries_resolve_to_classes, test_op_class_exists,
  test_forward_input_names_match_manifest,
  test_forward_second_arg_matches_manifest,
  test_init_includes_manifest_params, test_construction_smoke
  (~96 nodes), all covered by C3/C4/C5. Rename the file to
  test_elementwise_binary_broadcast.py to reflect what remains: the
  bidirectional-broadcast behavior smoke vs. PyTorch reference.

Net: ~287 fewer pytest nodes on testbed; broadcast behavior coverage
preserved.
