+++ tests/ops/test_elementwise_binary_broadcast.py
@@ -1,0 +1,157 @@
+"""Bidirectional-broadcast behavior tests for ``elementwise_binary`` ops.
+
+L1 signature/parity (forward arg names, ``__init__`` defaults, manifest
+entry resolution, construction smoke) is enforced by
+``scripts/validate_manifest.py`` strict-parity gates C3/C4/C5; per-op
+pytest fanouts mirroring the manifest are redundant. What remains here
+is the load-bearing external behavior: bidirectional broadcast against
+a PyTorch reference.
+"""
+
+from __future__ import annotations
+
+import pytest
+import torch
+
+import tileops.ops.elementwise as elementwise_mod
+
+
+# (op_name, dtype, gen_a, gen_b, ref_fn) — float-kernel paths.
+_BROADCAST_FLOAT_OPS = [
+    ("AddFwdOp",
+     torch.float16,
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda a, b: a + b),
+    ("SubFwdOp",
+     torch.float16,
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda a, b: a - b),
+    ("MulFwdOp",
+     torch.float16,
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda a, b: a * b),
+    ("DivFwdOp",
+     torch.float16,
+     lambda s, d: torch.rand(*s, dtype=d, device="cuda") + 0.1,
+     lambda s, d: torch.rand(*s, dtype=d, device="cuda") + 0.1,
+     lambda a, b: a / b),
+    ("RemainderFwdOp",
+     torch.float16,
+     lambda s, d: torch.rand(*s, dtype=d, device="cuda") + 0.1,
+     lambda s, d: torch.rand(*s, dtype=d, device="cuda") + 0.1,
+     lambda a, b: torch.remainder(a, b)),
+    ("PowFwdOp",
+     torch.float16,
+     lambda s, d: torch.rand(*s, dtype=d, device="cuda") + 0.5,
+     lambda s, d: torch.rand(*s, dtype=d, device="cuda") * 2.0,
+     lambda a, b: torch.pow(a, b)),
+    ("FloorDivideFwdOp",
+     torch.float16,
+     lambda s, d: torch.rand(*s, dtype=d, device="cuda") + 0.1,
+     lambda s, d: torch.rand(*s, dtype=d, device="cuda") + 0.1,
+     lambda a, b: torch.floor_divide(a, b)),
+    ("LerpFwdOp",
+     torch.float16,
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda a, b: torch.lerp(a, b, 0.5)),
+    ("MaximumFwdOp",
+     torch.float16,
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda a, b: torch.maximum(a, b)),
+    ("MinimumFwdOp",
+     torch.float16,
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda a, b: torch.minimum(a, b)),
+    ("EqFwdOp",
+     torch.float16,
+     lambda s, d: (torch.randn(*s, dtype=d, device="cuda") > 0).to(d),
+     lambda s, d: (torch.randn(*s, dtype=d, device="cuda") > 0).to(d),
+     lambda a, b: a == b),
+    ("NeFwdOp",
+     torch.float16,
+     lambda s, d: (torch.randn(*s, dtype=d, device="cuda") > 0).to(d),
+     lambda s, d: (torch.randn(*s, dtype=d, device="cuda") > 0).to(d),
+     lambda a, b: a != b),
+    ("GtFwdOp",
+     torch.float16,
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda a, b: a > b),
+    ("LtFwdOp",
+     torch.float16,
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda a, b: a < b),
+    ("GeFwdOp",
+     torch.float16,
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda a, b: a >= b),
+    ("LeFwdOp",
+     torch.float16,
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda s, d: torch.randn(*s, dtype=d, device="cuda"),
+     lambda a, b: a <= b),
+    ("LogicalAndFwdOp",
+     torch.float16,
+     lambda s, d: (torch.randn(*s, dtype=d, device="cuda") > 0).to(d),
+     lambda s, d: (torch.randn(*s, dtype=d, device="cuda") > 0).to(d),
+     lambda a, b: torch.logical_and(a, b)),
+    ("LogicalOrFwdOp",
+     torch.float16,
+     lambda s, d: (torch.randn(*s, dtype=d, device="cuda") > 0).to(d),
+     lambda s, d: (torch.randn(*s, dtype=d, device="cuda") > 0).to(d),
+     lambda a, b: torch.logical_or(a, b)),
+]
+
+_BROADCAST_INT_OPS = [
+    ("BitwiseAndFwdOp",
+     torch.int32,
+     lambda s, d: torch.randint(-1000, 1000, s, dtype=d, device="cuda"),
+     lambda s, d: torch.randint(-1000, 1000, s, dtype=d, device="cuda"),
+     lambda a, b: torch.bitwise_and(a, b)),
+    ("BitwiseOrFwdOp",
+     torch.int32,
+     lambda s, d: torch.randint(-1000, 1000, s, dtype=d, device="cuda"),
+     lambda s, d: torch.randint(-1000, 1000, s, dtype=d, device="cuda"),
+     lambda a, b: torch.bitwise_or(a, b)),
+    ("BitwiseXorFwdOp",
+     torch.int32,
+     lambda s, d: torch.randint(-1000, 1000, s, dtype=d, device="cuda"),
+     lambda s, d: torch.randint(-1000, 1000, s, dtype=d, device="cuda"),
+     lambda a, b: torch.bitwise_xor(a, b)),
+]
+
+
+@pytest.mark.smoke
+@pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA required")
+@pytest.mark.parametrize(
+    "op_name, dtype, gen_a, gen_b, ref_fn",
+    _BROADCAST_FLOAT_OPS + _BROADCAST_INT_OPS,
+    ids=lambda v: v if isinstance(v, str) else None,
+)
+def test_binary_op_bidirectional_broadcast(
+    op_name: str, dtype: torch.dtype, gen_a, gen_b, ref_fn,
+) -> None:
+    """Bidirectional broadcast: (3,1) x (1,4) -> (3,4)."""
+    cls = getattr(elementwise_mod, op_name)
+    a_shape = (3, 1)
+    b_shape = (1, 4)
+    a = gen_a(a_shape, dtype)
+    b = gen_b(b_shape, dtype)
+    op = cls(a_shape=a_shape, b_shape=b_shape, dtype=dtype)
+    out = op(a, b)
+    ref = ref_fn(a, b)
+    assert tuple(out.shape) == (3, 4), (
+        f"{op_name}: expected output shape (3, 4), got {tuple(out.shape)}"
+    )
+    if out.dtype.is_floating_point:
+        torch.testing.assert_close(out, ref, atol=1e-2, rtol=1e-2)
+    else:
+        torch.testing.assert_close(out, ref.to(out.dtype))
